pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "An openADR VTN implementation built on top of Django web framework."
authors = ["Gr3at <gr3atdk@gmail.com>"]
readme = "README.md"


[tool.poetry.scripts]
make = 'scripts.poetry_scripts:django_makemigrations_and_migrate'
serve = 'scripts.poetry_scripts:django_runserver'


[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
lxml = "^5.1.0"
xmltodict = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
